:recycle: Refactor: Compare Submodule
